Added MantisBTHelper basic class, with simple function as POC.
This new helper only have a simple function to serve as POC. Related
Unit and Integration tests also added. A bug in the MantisBT Docker
container used prevent IT tests to be ran with Travis, tests are run
manually for now and will be set-up automatically once bug corrected.

Some impacts on existing code:
- added test files (src/test/resources/mantisbt) and updated testng.xml
- Updated defaut ServerContext port to -1 (impact on SFTP tests)