Added CMISHelper infrastructure, support AtomPub and WebService binding.
This new Helper uses Apache Chemistry. The Helper revolve around the
SessionHandler which will create CMIS sessions depending on the Context
provided (which may be using Atom or WebService binding)

Changes:
- added Chemistry dependencies to pom.xml
- created Helper and Context classes
- Added corresponding tests

Tests are using the public Alfresco CMIS repository available for now. A
Docker container may be added in the future to support local-only
testing.